+++ lib/widgets/refresh_indicator.dart
@@ -2,6 +2,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_refresh_indicator/flutter_refresh_indicator.dart';
 
+/// Defines the visual and behavioral style of the refresh indicator.
 enum RefreshIndicatorType { clamping, bouncing }
 
 /// Signature for the alias that is representing of the [RefreshIndicator] widget.
@@ -15,10 +16,12 @@ class RefreshIndicator extends StatefulWidget {
   const RefreshIndicator({
     super.key,
     this.type,
+    this.enabled = true,
     required this.onRefresh,
     required this.child,
   });
 
+  /// The visual and behavioral style of this refresh indicator.
   final RefreshIndicatorType? type;
 
   /// The callback that's called when the user has dragged the refresh indicator
@@ -26,6 +29,9 @@ class RefreshIndicator extends StatefulWidget {
   ///
   /// The returned [Future] must complete when the refresh operation is finished.
   final AsyncCallback onRefresh;
+
+  /// Whether users can pull to trigger the refresh operation.
+  final bool enabled;
 
   /// The widget to be contained as descendant by this widget.
   final Widget child;
@@ -43,19 +49,19 @@ class _RefreshIndicatorState extends State<RefreshIndicator> {
 
   @override
   Widget build(BuildContext context) {
-    final RefreshIndicatorType current = widget.type ??
-        (isBouncing
-            ? RefreshIndicatorType.bouncing
-            : RefreshIndicatorType.clamping);
+    final RefreshIndicatorType current =
+        widget.type ?? (isBouncing ? RefreshIndicatorType.bouncing : RefreshIndicatorType.clamping);
 
     if (current == RefreshIndicatorType.clamping) {
       return ClampingRefreshIndicator(
         onRefresh: widget.onRefresh,
+        enabled: widget.enabled,
         child: widget.child,
       );
     } else {
       return BouncingRefreshIndicator(
         onRefresh: widget.onRefresh,
+        enabled: widget.enabled,
         child: widget.child,
       );
     }